requirements.txt
# Core
python-dotenv>=1.0.0
pandas>=2.0.3
numpy>=1.24.3
scipy>=1.11.0

# ML
scikit-learn>=1.3.0
xgboost>=2.0.0
joblib>=1.3.0

# Data
yfinance>=0.2.33
twelvedata>=0.3.0
alpha-vantage>=2.3.1
finnhub-python>=2.4.18
requests>=2.31.0
websocket-client>=1.9.0
orjson>=3.11.7
ujson>=5.11.0

# Web / Dashboard
flask>=2.3.0
flask-cors>=4.0.0
flask-socketio>=5.3.0
eventlet>=0.33.0
python-multipart>=0.0.22

# Telegram
python-telegram-bot>=20.0
telethon>=1.42.0
cryptg>=0.5.2
pyaes>=1.6.1
rsa>=4.9.1

# Database (optional)
psycopg2-binary>=2.9.11
sqlalchemy>=2.0.48

# NLP / Sentiment
textblob>=0.17.1
feedparser>=6.0.12
beautifulsoup4>=4.14.3

# Social
tweepy>=4.16.0

# Scheduling
schedule>=1.2.0

# Monitoring
psutil>=5.9.0

# Serialization
orjson>=3.11.0

# Utilities
tabulate>=0.9.0
colorama>=0.4.6
aiofiles>=25.1.0